Project/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
qrcode==7.4.2
python-barcode==0.15.1
segno==1.6.0
reportlab==4.0.7
python-dotenv==1.0.0
treepoem
